Revert "changes made by MK"
